revert to floats as it works better(?)

+++ GroteOpdrachtV2/NeighbourSpace.cs
@@ -1,8 +1,8 @@
 ﻿namespace GroteOpdrachtV2 {
     public struct ValuePerNeighbour {
-        public double value;
+        public float value;
         public NeighbourSpace type;
-        public ValuePerNeighbour(double value, NeighbourSpace type) {
+        public ValuePerNeighbour(float value, NeighbourSpace type) {
             this.value = value;
             this.type = type;
         }
